管理员登录，超级管理员区分，产品、分类、用户、普通管理员操作实现和优化

+++ MyShopDesign/MyShopDesign/Areas/Admin/Controllers/CategoryController.cs
@@ -15,6 +15,11 @@ namespace MyShopDesign.Areas.Admin.Controllers
 
         public ActionResult Index(string searchName, string orderField = "Id desc", int pageIndex = 1)
         {
+            if (Session["person"] == null)
+            {
+                return Redirect("/Admin/AdminLogin/Index");
+            }
+
             IEnumerable<T_Shop_ProductCategory> query = db.T_Shop_ProductCategory;
 
             if (string.IsNullOrEmpty(searchName))
@@ -53,38 +58,62 @@ namespace MyShopDesign.Areas.Admin.Controllers
 
             List<T_Shop_ProductCategory> list = query.ToList();
             ViewBag.list = list;
+            ViewBag.search = searchName;
             return View();
         }
 
         [HttpGet]
         public ActionResult Add()
         {
+            if (Session["person"] == null)
+            {
+                return Redirect("/Admin/AdminLogin/Index");
+            }
+
             return View();
         }
 
         [HttpPost]
-        public ActionResult Add(T_Shop_ProductCategory category)
+        public int Add(T_Shop_ProductCategory category)
         {
+            IEnumerable<T_Shop_ProductCategory> p = db.T_Shop_ProductCategory.Where(m => m.Name==category.Name);
+            if(p.Count() != 0)
+            {
+                return 0;
+            }
+
             T_Shop_ProductCategory temp = new T_Shop_ProductCategory();
             temp.Name = category.Name;
             temp.Description = category.Description;
             db.T_Shop_ProductCategory.Add(temp);
             db.SaveChanges();
 
-            return RedirectToAction("index");
+            return 1;
         }
 
-        public ActionResult Delete(int id)
+        [HttpPost]
+        public int Delete(int id)
         {
+            IEnumerable<T_Shop_Product>  p = db.T_Shop_Product.Where(m => m.CategoryId == id);
+            if (p.Count() != 0)
+            {
+                return 0;
+            }
+
             T_Shop_ProductCategory temp = db.T_Shop_ProductCategory.Find(id);
             db.T_Shop_ProductCategory.Remove(temp);
             db.SaveChanges();
-            return RedirectToAction("index");
+            return 1;
         }
 
         [HttpGet]
         public ActionResult Edit(int id)
         {
+            if (Session["person"] == null)
+            {
+                return Redirect("/Admin/AdminLogin/Index");
+            }
+
             T_Shop_ProductCategory temp = db.T_Shop_ProductCategory.Find(id);
             ViewBag.item = temp;
             return View();
@@ -100,5 +129,6 @@ namespace MyShopDesign.Areas.Admin.Controllers
             db.SaveChanges();
             return RedirectToAction("index");
         }
+
     }
 }
